+++ src/matyan/registry.py
@@ -1,3 +1,5 @@
+from typing import Dict, Type
+
 __author__ = 'Artur Barseghyan'
 __copyright__ = '2013-2019 Artur Barseghyan'
 __license__ = 'MPL-1.1 OR GPL-2.0-only OR LGPL-2.0-or-later'
@@ -7,8 +9,9 @@ __all__ = (
 
 
 class Registry(type):
+    """Base registry."""
 
-    REGISTRY = {}
+    REGISTRY: Dict[str, Type]
 
     def __new__(cls, name, bases, attrs):
         new_cls = type.__new__(cls, name, bases, attrs)