+++ src/matyan/renderers/rest.py
@@ -1,4 +1,5 @@
 from typing import Dict
+from textwrap import indent
 
 from ..labels import BRANCH_TYPES, BRANCH_TYPE_OTHER
 from .base import BaseRenderer
@@ -7,38 +8,45 @@ __author__ = 'Artur Barseghyan'
 __copyright__ = '2019 Artur Barseghyan'
 __license__ = 'GPL-2.0-only OR LGPL-2.0-or-later'
 __all__ = (
-    'MarkdownRenderer',
-    'HistoricalMarkdownRenderer',
+    'RestructuredTextRenderer',
 )
 
 
-class MarkdownRenderer(BaseRenderer):
+class RestructuredTextRenderer(BaseRenderer):
 
-    uid: str = 'markdown'
+    uid: str = 'rest'
 
     def append_release(self, release_label: str):
-        self.changelog.append("\n### {}".format(release_label))
+        self.changelog.append(
+            "\n{}\n{}".format(
+                release_label,
+                "-" * len(release_label)
+            )
+        )
 
     def append_feature_type(self, branch_type: str):
+        branch_type_label = BRANCH_TYPES.get(branch_type)
         self.changelog.append(
-            "\n#### {}{}".format(
-                BRANCH_TYPES.get(branch_type),
+            "\n{}\n{}{}".format(
+                branch_type_label,
+                "~" * len(branch_type_label),
                 '\n' if branch_type == BRANCH_TYPE_OTHER else ''
             )
         )
 
     def append_ticket_title(self, ticket_number: str, ticket_title: str):
+        ticket_title_label = "{} {}".format(ticket_number, ticket_title)
         self.changelog.append(
-            "\n##### {} {}".format(
-                ticket_number,
-                ticket_title
+            "\n{}\n{}".format(
+                ticket_title_label,
+                "^" * len(ticket_title_label)
             )
         )
 
     def append_ticket_description(self, ticket_description: str):
         self.changelog.append(
-            "\n```\n{}\n```".format(
-                ticket_description.strip()
+            "\n.. code-block: text\n\n{}".format(
+                indent(ticket_description.strip(), "    ")
             )
         )
 
@@ -55,29 +63,5 @@ class MarkdownRenderer(BaseRenderer):
                 ),
                 commit_data['title'],
                 commit_data['author']
-            )
-        )
-
-
-class HistoricalMarkdownRenderer(MarkdownRenderer):
-
-    uid: str = 'historical-markdown'
-
-    def append_release(self, release_label: str):
-        self.changelog.append("\n### {}".format(release_label))
-
-    def append_feature_type(self, branch_type: str):
-        self.changelog.append(
-            "\n**{}**{}".format(
-                BRANCH_TYPES.get(branch_type),
-                '\n' if branch_type == BRANCH_TYPE_OTHER else ''
-            )
-        )
-
-    def append_ticket_title(self, ticket_number: str, ticket_title: str):
-        self.changelog.append(
-            "\n*{} {}*".format(
-                ticket_number,
-                ticket_title
             )
         )